@@ -1,16 +1,15 @@
 name: estados_municipios
 description: Pacote para facilitar e agilizar o processo de escolha de municipios e estados brasileiros. Todas as informações são tiradas da API do IBGE.
-version: 1.0.1+2
+version: 2.0.0-nullsafety.0
 homepage: https://github.com/rodrigobastosv/estados_municipios
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
-  flutter: ">=1.17.0 <2.0.0"
+  sdk: '>=2.12.0-0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  http: ^0.12.2
+  http: ^0.13.0-nullsafety.0
 
 dev_dependencies:
   flutter_test:
