@@ -2,10 +2,17 @@ import 'regiao.dart';
 
 /// Classe que modela um Estado na API do IBGE
 class Estado {
-  int id;
-  String sigla;
-  String nome;
-  Regiao regiao;
+  /// Id do estado IBGE
+  int? id;
+
+  /// Sigla identificadora do estado
+  String? sigla;
+
+  /// Nome do estado
+  String? nome;
+
+  /// Região do estado
+  Regiao? regiao;
 
   Estado({
     this.id,
@@ -14,6 +21,7 @@ class Estado {
     this.regiao,
   });
 
+  /// Método para desserializar o objeto
   Estado.fromJson(Map<String, dynamic> json) {
     id = json['id'];
     sigla = json['sigla'];
@@ -21,13 +29,14 @@ class Estado {
     regiao = json['regiao'] != null ? Regiao.fromJson(json['regiao']) : null;
   }
 
+  /// Método para serializar o objeto
   Map<String, dynamic> toJson() {
-    final data = {};
+    final data = <String, dynamic>{};
     data['id'] = id;
     data['sigla'] = sigla;
     data['nome'] = nome;
     if (regiao != null) {
-      data['regiao'] = regiao.toJson();
+      data['regiao'] = regiao!.toJson();
     }
     return data;
   }
