@@ -2,12 +2,17 @@ import 'microrregiao.dart';
 
 /// Classe que modela um Município na API do IBGE
 class Municipio {
-  int id;
-  String nome;
-  Microrregiao microrregiao;
+  int? id;
+  String? nome;
+  Microrregiao? microrregiao;
 
-  Municipio({this.id, this.nome, this.microrregiao});
+  Municipio({
+    this.id,
+    this.nome,
+    this.microrregiao,
+  });
 
+  /// Método para desserializar o objeto
   Municipio.fromJson(Map<String, dynamic> json) {
     id = json['id'];
     nome = json['nome'];
@@ -16,12 +21,13 @@ class Municipio {
         : null;
   }
 
+  /// Método para serializar o objeto
   Map<String, dynamic> toJson() {
-    final data = {};
+    final data = <String, dynamic>{};
     data['id'] = id;
     data['nome'] = nome;
     if (microrregiao != null) {
-      data['microrregiao'] = microrregiao.toJson();
+      data['microrregiao'] = microrregiao!.toJson();
     }
     return data;
   }
