@@ -1,8 +1,8 @@
 /// Classe que modela uma Região na API do IBGE
 class Regiao {
-  int id;
-  String sigla;
-  String nome;
+  int? id;
+  String? sigla;
+  String? nome;
 
   Regiao({
     this.id,
@@ -10,14 +10,16 @@ class Regiao {
     this.nome,
   });
 
+  /// Método para deserializar o objeto
   Regiao.fromJson(Map<String, dynamic> json) {
     id = json['id'];
     sigla = json['sigla'];
     nome = json['nome'];
   }
 
+  /// Método para serializar o objeto
   Map<String, dynamic> toJson() {
-    final data = {};
+    final data = <String, dynamic>{};
     data['id'] = id;
     data['sigla'] = sigla;
     data['nome'] = nome;
